move navbar to molecules
